requirements.txt
anyio==4.12.1
attrs==25.4.0
beautifulsoup4==4.14.3
blinker==1.9.0
bs4==0.0.2
build==1.4.0
CacheControl==0.14.4
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
cleo==2.1.0
click==8.3.1
crashtest==0.4.1
cryptography==46.0.5
data_url==1.3.1
distlib==0.4.0
dulwich==1.1.0
fastjsonschema==2.21.2
filelock==3.24.3
findpython==0.7.1
Flask==3.1.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
installer==0.7.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.1.0
jaraco.functools==4.4.0
jeepney==0.9.0
Jinja2==3.1.6
keyring==25.7.0
MarkupSafe==3.0.3
more-itertools==10.8.0
msgpack==1.1.2
outcome==1.3.0.post0
packaging==26.0
pbs-installer==2026.2.11
pillow==12.1.1
pipenv==2026.0.3
pkginfo==1.12.1.2
platformdirs==4.9.2
poetry==2.3.2
poetry-core==2.3.1
pycparser==3.0
pyproject_hooks==1.2.0
PySocks==1.7.1
RapidFuzz==3.14.3
requests==2.32.5
requests-toolbelt==1.0.0
SecretStorage==3.5.0
selenium==4.41.0
setuptools==82.0.1
shellingham==1.5.4
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8.3
tomlkit==0.14.0
trio==0.33.0
trio-websocket==0.12.2
trove-classifiers==2026.1.14.14
typing_extensions==4.15.0
urllib3==2.6.3
virtualenv==20.38.0
websocket-client==1.9.0
Werkzeug==3.1.6
wsproto==1.3.2
zstandard==0.25.0
